+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 - Lazy loading implemented for bigger libraries (Plotly, jsPDF) to reduce initial bundle size
 - Dynamically import components as necessary to chunk the application further
-- User defined paramaters are now highlighted when viewing Computation parameters
+- User defined parameters are now highlighted when viewing Computation parameters
 - Demo computations are now displayed with a badge in the list ([#441](https://gitlab.heigit.org/climate-action/web-app/-/issues/441))
+- Ability to toggle visibility of vector layer categories via the legend ([#162](https://gitlab.heigit.org/climate-action/web-app/-/issues/162))
 
 ### Fixed